backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
azure-storage-blob
python-multipart
